move data outside oot submodule
