+++ fast64_internal/constants.py
@@ -3,7 +3,7 @@ from typing import Optional
 
 class GameData:
     def __init__(self, game_editor_mode: Optional[str] = None):
-        from .oot.data import OoT_Data
+        from .data.oot import OoT_Data
 
         self.z64 = OoT_Data()
