+++ fast64_internal/data/oot/__init__.py
