+++ fast64_internal/data/oot/oot_data.py
