+++ fast64_internal/data/oot/oot_enum_data.py
