+++ fast64_internal/data/oot/oot_getters.py
