+++ fast64_internal/data/oot/oot_object_data.py
