+++ fast64_internal/data/oot/xml/ActorList.xml
