+++ fast64_internal/data/oot/xml/EnumData.xml
