+++ fast64_internal/data/oot/xml/ObjectList.xml
